componentiza e add filtro para stack
